rename tests level1 to CDBStructure
